% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scagnostics.r
\name{print.scagdf}
\alias{print.scagdf}
\title{Print scagnostics data structure}
\usage{
\method{print}{scagdf}(x, ...)
}
\arguments{
\item{x}{Object to be printed}

\item{...}{Extra arguments}
}
\description{
Print scagnostics data structure
}
